requirements.txt
langchain_core==1.1.0
langchain_google_genai==3.2.0
langgraph==1.0.4
python-dotenv==1.2.1
